Refactor patch generation
