+++ source/satellite_dataset/typing.py
@@ -1,5 +1,7 @@
 from typing import TypedDict
 
+from numpy import ndarray
+
 
 class DownloadConfig(TypedDict):
     archive: str
@@ -10,3 +12,12 @@ class DownloadConfig(TypedDict):
 class SatelliteDataset(TypedDict):
     path: str
     archive: str
+
+
+class ImgGeoDataArrays(TypedDict):
+    image: ndarray
+    latitude: ndarray
+    longitude: ndarray
+    local_time: ndarray
+    incidence_angle: ndarray
+    emission_angle: ndarray
